+++ multiboot2/Changelog.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 - MSRV is 1.56.1
+- **BREAKING** fixed lifetime issues: `VBEInfoTag` is no longer static
+- fixed another internal lifetime issue
+- `BootInformation::framebuffer_tag()` now returns
+  `Option<Result<FramebufferTag, UnknownFramebufferType>>` instead of
+  `Option<FramebufferTag>` which prevents a possible panic. If the `--unstable`
+  feature is used, `UnknownFramebufferType` implements `core::error::Error`.
 
 ## 0.14.2 (2023-03-17)
 - documentation fixes